fix(): rename ProductCard.component and Products.component and fix deleteProductCache
